requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==1.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
docker-pycreds==0.4.0
et-xmlfile==2.0.0
Farama-Notifications==0.0.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.67.0
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.6.0
importlib-metadata==4.13.0
jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.19
lazy-loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
openai==1.69.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opensimplex==0.4.5.1
opt-einsum==3.4.0
optree==0.13.1
orjson==3.10.16
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
propcache==0.3.1
protobuf==5.28.2
psutil==6.1.0
pydantic==2.11.0
pydantic-core==2.33.0
pydantic-settings==2.8.1
pygame==2.6.1
pyglet==1.5.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
sb3-contrib==2.3.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==4.0.1
sentry-sdk==2.17.0
setproctitle==1.3.3
smmap==5.0.1
sniffio==1.3.1
sqlalchemy==2.0.40
stable-baselines3==1.7.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf-keras==2.19.0
threadpoolctl==3.6.0
tifffile==2024.9.20
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.5.1
tqdm==4.66.6
transformers==4.50.2
triton==3.1.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.5
werkzeug==3.0.4
wrapt==1.17.0
yarl==1.18.3
zipp==3.20.2
zstandard==0.23.0
